codigo_com_ig/requirements.txt
tk
geopy
requests
reportlab
folium
pywebview
